+++ src/shared/Router.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Layout from '../components/Layout';
-import Login from "../pages/Login";
-import Join from "../pages/Join";
+import Login from "../components/Login";
+import Join from "../components/Join";
 
 const Router = () => {
   return (
